Move `DnsResolverTest` tests (which require UDP sockets) to common
