+++ ktor-client/ktor-client-cio/common/test/CioDnsResolverTest.kt
@@ -6,30 +6,26 @@ package io.ktor.client.engine.cio
 
 import io.ktor.network.selector.*
 import io.ktor.network.sockets.*
+import io.ktor.test.*
 import io.ktor.utils.io.core.buildPacket
 import kotlinx.coroutines.*
+import kotlinx.coroutines.test.TestResult
 import kotlinx.io.*
-import java.net.UnknownHostException
-import kotlin.random.Random
 import kotlin.test.*
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
-import kotlin.time.measureTime
+import kotlin.use
 
-class DnsResolverTest {
+class CioDnsResolverTest {
 
-    @Test
-    fun testJvmDnsResolverResolvesLocalhost() = runBlocking {
-        val ips = JvmDnsResolver().invoke("localhost")
-        assertTrue(ips.isNotEmpty(), "localhost must resolve to at least one address")
-        assertTrue(
-            ips.any { it == "127.0.0.1" || it == "0:0:0:0:0:0:0:1" || it == "::1" },
-            "expected loopback in $ips",
-        )
+    private fun runTestWithNoDelaySkipping(testBody: suspend CoroutineScope.() -> Unit): TestResult = runTest {
+        withContext(Dispatchers.Default.limitedParallelism(1)) {
+            testBody()
+        }
     }
 
     @Test
-    fun testCioDnsResolverParsesAResponse() = runBlocking {
+    fun testCioDnsResolverParsesAResponse() = runTestWithNoDelaySkipping {
         withFakeDnsServer { server, port ->
             val resolved = async {
                 CioDnsResolver(server = "127.0.0.1", port = port, timeout = 2.seconds)
@@ -42,7 +38,7 @@ class DnsResolverTest {
     }
 
     @Test
-    fun testCioDnsResolverReturnsAaaaOnlyForIpv6OnlyHost() = runBlocking {
+    fun testCioDnsResolverReturnsAaaaOnlyForIpv6OnlyHost() = runTestWithNoDelaySkipping {
         withFakeDnsServer { server, port ->
             val resolved = async {
                 CioDnsResolver(server = "127.0.0.1", port = port, timeout = 2.seconds)
@@ -55,7 +51,7 @@ class DnsResolverTest {
     }
 
     @Test
-    fun testCioDnsResolverReturnsEmptyOnNxDomain() = runBlocking {
+    fun testCioDnsResolverReturnsEmptyOnNxDomain() = runTestWithNoDelaySkipping {
         withFakeDnsServer { server, port ->
             val resolved = async {
                 CioDnsResolver(server = "127.0.0.1", port = port, timeout = 2.seconds)
@@ -68,7 +64,7 @@ class DnsResolverTest {
     }
 
     @Test
-    fun testCioDnsResolverHappyEyeballsPrefersIpv6() = runBlocking {
+    fun testCioDnsResolverHappyEyeballsPrefersIpv6() = runTestWithNoDelaySkipping {
         withFakeDnsServer { server, port ->
             val resolved = async {
                 CioDnsResolver(server = "127.0.0.1", port = port, timeout = 2.seconds)
@@ -81,7 +77,7 @@ class DnsResolverTest {
     }
 
     @Test
-    fun testCioDnsResolverThrowsWhenAllQueriesTimeOut() = runBlocking<Unit> {
+    fun testCioDnsResolverThrowsWhenAllQueriesTimeOut() = runTestWithNoDelaySkipping {
         SelectorManager().use { selector ->
             aSocket(selector).udp().bind(InetSocketAddress("127.0.0.1", 0)).use { server ->
                 val port = (server.localAddress as InetSocketAddress).port
@@ -97,7 +93,7 @@ class DnsResolverTest {
     }
 
     @Test
-    fun testCioDnsResolverRejectsEmptyHostname() = runBlocking<Unit> {
+    fun testCioDnsResolverRejectsEmptyHostname() = runTestWithNoDelaySkipping {
         val failure = assertFailsWith<IllegalArgumentException> {
             CioDnsResolver(server = "127.0.0.1").invoke("")
         }
@@ -105,7 +101,7 @@ class DnsResolverTest {
     }
 
     @Test
-    fun testCioDnsResolverAcceptsFqdnWithTrailingDot() = runBlocking {
+    fun testCioDnsResolverAcceptsFqdnWithTrailingDot() = runTestWithNoDelaySkipping {
         withFakeDnsServer { server, port ->
             val resolved = async {
                 CioDnsResolver(server = "127.0.0.1", port = port, timeout = 2.seconds)
@@ -115,27 +111,6 @@ class DnsResolverTest {
             server.replyEmpty()
             assertEquals(listOf("203.0.113.99"), resolved.await())
         }
-    }
-
-    @Test
-    fun testJvmDnsResolverIsCooperativelyCancellable() = runBlocking<Unit> {
-        // Smoke test only. A strong test for `runInterruptible` interrupting an in-flight blocking
-        // `InetAddress.getAllByName` call would need a hostname whose DNS lookup blocks for tens of
-        // seconds, which is not portable across CI environments. This test merely verifies that the
-        // resolver returns control on cancellation within a generous bound.
-        val resolver = JvmDnsResolver()
-        val job = launch {
-            while (isActive) {
-                try {
-                    resolver.invoke("nonexistent-${Random.nextInt()}.invalid")
-                } catch (_: UnknownHostException) {
-                    // expected — `.invalid` TLD is reserved as guaranteed-non-resolving (RFC 6761)
-                }
-            }
-        }
-        delay(50)
-        val elapsed = measureTime { job.cancelAndJoin() }
-        assertTrue(elapsed < 5.seconds, "cancellation took $elapsed; expected runInterruptible to bound it")
     }
 
     @Test
